requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.1
Django==5.2.6
django-cors-headers==4.9.0
django-oauth-toolkit==3.0.1
django-safedelete==1.4.1
djangorestframework==3.16.1
dnspython==2.8.0
drf-yasg==1.21.11
email-validator==2.3.0
firebase_admin==7.1.0
google-api-core==2.25.1
google-api-python-client==2.182.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.4.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.75.0
grpcio-status==1.75.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
inflection==0.5.1
jwcrypto==1.5.6
msgpack==1.1.1
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
proto-plus==1.26.1
protobuf==6.32.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.4
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
